fix failure to start Read on newly installed system, dev.laptop.org #10218
Read from USB stick using object chooser fails, only on the first
attempt on a newly installed system.

The error we are seeing is that the '(env.get_profile_path(),
'data'))' does not exist yet at that time. It gets created when an
activity like Browse or Terminal has been run for the first time or
you start Read by resuming the file on the external device. The issue
is not read specific, same is true for Memorize when you use the
object chooser in the create tab and try to open an image from the
external device.

Tested on os850 on XO-1.5.

Reviewed-by: James Cameron <quozl@laptop.org>
Reviewed-by: Daniel Drake <dsd@laptop.org>
Tested-by: James Cameron <quozl@laptop.org>