Restructure for new /usr/bin/sugar-activity behaviour
This patch accompanies a sugar patch titled
"sugar-activity: make independent of sugar-toolkit GTK versions"

The core Activity-instantiating functionality of main.py has been moved
into the sugar-activity binary and can be removed from here.

The remaining functionality (and everything that is GTK-specific) is moved
into the Activity class in this commit.

This is needed to make /usr/bin/sugar-activity independent of sugar/sugar3
and GTK2/GTK3, which is a crucial step for GTK3 activity support.

Signed-off-by: Daniel Drake <dsd@laptop.org>
Acked-by: Sascha Silbe <silbe@activitycentral.com>