icon: use GDK for pixbuf-to-cairo conversion
We currently use some strange code inside hippo to paint a pixbuf onto
a cairo surface. This can be more effectively done within GDK.

This solves a Sugar crash-on-startup on Fedora 17, where this hippo
code is crashing, and uses the same underlying mechanism as
sugar-toolkit-gtk3.

Also use cairo.Context.set_source_rgb() for setting the background
colour, which avoids the requirement to encapsulate the cairo context
at an early stage (which would make the rest of this patch a little
more complicated).

Signed-off-by: Daniel Drake <dsd@laptop.org>
Acked-by:  Simon Schampijer <simon@schampijer.de>