Add DescriptionEntry to the activity sub-toolbar
This is the implementation of the 'Write to Journal anytime'
feature [1].

The patch itself adds a DescriptionItem to the activity
sub-toolbar to make editing a Journal entry description
from within the activity possible. The code has the same
error handling as the TitleEntry.

The patch is an adaptation from the one that landed in
sugar-toolkit-gtk3 and has been ported to the
sugar-toolkit. After talking with teachers and the learning team,
I found out that they are very excited about this feature.
Doing this change not only in sugar-toolkit-gtk3 and limit it
to the activities that have been ported to the new toolkit
we will have a more consistent experience across the activities
in the 0.96 release.

The original patch was done by Simon Schampijer <simon@laptop.org>
and Walter Bender <walter.bender@gmail.com>

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>

[1] http://wiki.sugarlabs.org/go/Features/Write_to_journal_anytime