Remove the Naming Alert
With the new 'Write to Journal anytime' feature it is possible
to write a description within an activity. Hopefully this leads
to more usage of the Journal for reflection by the learner. The
hinting nature ('you have not named your session yet') of the
Naming Alert has not been replaced, though.

The patch is an adaptation from the one that landed in
sugar-toolkit-gtk3 and has been ported to the
sugar-toolkit. After talking with teachers and the learning team,
I found out that they are very excited about this feature.
Doing this change not only in sugar-toolkit-gtk3 and limit it
to the activities that have been ported to the new toolkit
we will have a more consistent experience across the activities
in the 0.96 release.

The original patch is from Simon Schampijer <simon@laptop.org>

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>