Save As - MouseDetector timeout_id need reset
Backport of sugar-toolkit-gtk3:fd15e76dabee72e7cf4d0369b63277353c3ae33f

The mouse detector have a variable to store the id for the timeout.
The stop() method check if the id is null but do not set the id to null
after remove the GObject timeout, then try to remove the timeout multiple times.

The efect can be seen if a user move the mouse slowly over the icons
in the Sugar home, in shell.log we see many lines like:

sugar/graphics/palettewindow.py:100: Warning: Source ID 2464 was not found
when attempting to remove it