Save As - show a save-as alert on stop
Backport of sugar-toolkit-gtk3:ba6993d70522524941fec6721d38ff6d79ec4a68

When an activity is to stop, and the save-as setting is enabled, and the
journal entry has not been named, show an alert asking for a name, with
keyboard focus in the name entry with the name selected.

Return key commits the save with the name as shown.

Escape key cancels the alert and the stop.

While the alert is shown, the activity stop button(s) will be
insensitive.

As a side-effect, when an entry is resumed, switching to the journal
will show a clone entry with the same name.  When the activity is
stopped, the clone entry will be removed or renamed.  Forced power down
and reboot will show the clone entry.

https://wiki.sugarlabs.org/go/Features/Save-As

Tested on Ubuntu 16.04 with GTK+ 2 HelloWorld 04fb9be (v3).

Signed-off-by: James Cameron <quozl@laptop.org>